feat: Implement `useChatInput` for `ChatContainer` component
